% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data51.R
\docType{data}
\name{data51}
\alias{data51}
\title{Data set number 1 for descriptive statistics}
\format{
\code{data51} is a data frame with 10 rows and 1 column:
\describe{
  \item{x}{Final score of each student}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data51
}
\description{
A dataset of the final scores of 10 students
}
\examples{
data51

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.122-123.
}
\keyword{datasets}
