% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data52.R
\docType{data}
\name{data52}
\alias{data52}
\title{Data set number 2 for descriptive statistics}
\format{
\code{data52} is a data frame with 5 rows and 2 columns:
\describe{
  \item{x}{Midterm score of English subject}
  \item{y}{Midterm score of Thai subject}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data52
}
\description{
A dataset of midterm scores from 5 students. Each student took the exam two subjects: Thai and English
}
\examples{
data52

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.125-128.
}
\keyword{datasets}
