% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data53.R
\docType{data}
\name{data53}
\alias{data53}
\title{Data set number 3 for descriptive statistics}
\format{
\code{data53} is a data frame with 10 rows and 1 column:
\describe{
  \item{x}{Weight of each person}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data53
}
\description{
A dataset of the weights of 10 persons measure in kilogram
}
\examples{
data53

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.129-130.
}
\keyword{datasets}
