% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data54.R
\docType{data}
\name{data54}
\alias{data54}
\title{Data set number 4 for descriptive statistics}
\format{
\code{data54} is a data frame with 5 rows and 3 columns:
\describe{
  \item{name}{Student name}
  \item{age}{Student age}
  \item{gender}{Student gender}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data54
}
\description{
A dataset of five students. Each student has name, age and gender
}
\examples{
data54

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.147.
}
\keyword{datasets}
