% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data81.R
\docType{data}
\name{data81}
\alias{data81}
\title{Data set for statistical hypothesis testing}
\format{
\code{data81} is a data frame with 10 rows and 2 columns:
\describe{
  \item{wtcar.a}{Weight of car type A}
  \item{wtcar.b}{Weight of car type B}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data81
}
\description{
A dataset of the weight of two car types (type A and type B). The weight measured in kilograms
}
\examples{
data81

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.237.
}
\keyword{datasets}
