% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data82.R
\docType{data}
\name{data82}
\alias{data82}
\title{Data set  for the paired t-test}
\format{
\code{data82} is a data frame with 15 rows and 2 columns:
\describe{
  \item{pretest}{The pre-test score}
  \item{posttest}{The post-test score}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data82
}
\description{
A dataset of the pre-test and post-test scores
}
\examples{
data82

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.247.
}
\keyword{datasets}
