% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data91.R
\docType{data}
\name{data91}
\alias{data91}
\title{Data set  for simple linear regression model}
\format{
\code{data91} is a data frame with 9 rows and 2 columns:
\describe{
  \item{Advcosts}{The advertising costs are measured in a hundred thousand baht}
  \item{Sales}{The sales are measured in a million bath}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data91
}
\description{
A dataset of the advertising costs and sales
}
\examples{
data91

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.265-266.
}
\keyword{datasets}
