% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data92.R
\docType{data}
\name{data92}
\alias{data92}
\title{Data set for multiple linear regression model}
\format{
\code{data91} is a data frame with 12 rows and 3 columns:
\describe{
  \item{x1}{The salary of employees is measured in baht}
  \item{x2}{The overtime of employees is measured in baht}
  \item{y}{The savings per month of employees are measured in baht}
}
}
\source{
<https://www.researchgate.net/publication/371944275_porkaermxarlaeakarprayuktchingan_R_and_its_applications>
}
\usage{
data92
}
\description{
A dataset of salary, overtime, and savings per month of 12 employees
}
\examples{
data92

}
\references{
Rattanalertnusorn, A. (2024). R and Its Application. TPN Press Ltd., Part. ISBN: 978-616-603-060-0. pp.275.
}
\keyword{datasets}
